refactor computer class
